Refactor store cross-cutting concerns
