+++ src/options/pages/brickEditor/EditPage.tsx
@@ -29,10 +29,10 @@ import BootstrapSwitchButton from "bootstrap-switch-button-react";
 import "./EditPage.scss";
 import { useDispatch, useSelector } from "react-redux";
 import { RawConfig } from "@/core";
-import { selectExtensions } from "@/options/selectors";
+import { selectExtensions } from "@/store/extensionsSelectors";
 import { useTitle } from "@/hooks/title";
 import { HotKeys } from "react-hotkeys";
-import { workshopSlice } from "@/options/slices";
+import workshopSlice from "@/store/workshopSlice";
 import useLogContext from "@/options/pages/brickEditor/useLogContext";
 import useFetch from "@/hooks/useFetch";
 import { loadBrickYaml } from "@/runtime/brickYaml";
